requirements.txt
distro==1.9.0
JPype1==1.5.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tabula-py==2.9.3
tzdata==2024.1
astroid==3.2.4
colorama==0.4.6
dill==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
isort==5.13.2
JPype1==1.5.0
mccabe==0.7.0
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
pylint==3.2.6
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tabula-py==2.9.3
tomlkit==0.13.2
tzdata==2024.1
